refactor: move SpinalSimLibTester
