+++ lib/src/test/scala/spinal/lib/Utils.scala
@@ -1,9 +1,13 @@
 package spinal.lib
 
 import spinal.core._
+import spinal.core.sim._
 import spinal.core.formal._
 import spinal.lib._
 import spinal.lib.formal._
+
+import org.scalatest.funsuite.AnyFunSuite
+import scala.util.Random
 
 class FormalHistoryModifyableTester extends SpinalFormalFunSuite {
   test("pop_any") {
@@ -88,5 +92,45 @@ class FormalHistoryModifyableTester extends SpinalFormalFunSuite {
         results.map(x => cover(x.fire))
         cover(results(0).fire && results(2).fire)
       })
+  }
+}
+
+class SpinalSimLibTester extends AnyFunSuite {
+  for (bitCount <- 0 until 12) {
+    test("CountOnes" + bitCount) {
+      LutInputs(Random.nextInt(5) + 2).on {
+        SimConfig.noOptimisation
+          .compile(new Component {
+            val input = in Bits (bitCount bits)
+            val output = out(CountOne(input))
+          })
+          .doSim(seed = 42) { dut =>
+            for (_ <- 0 until 100 + (1 << bitCount) * 4) {
+              dut.input.randomize()
+              sleep(1)
+              assert(dut.output.toInt === dut.input.toBigInt.bitCount)
+            }
+          }
+      }
+    }
+  }
+
+  for (bitCount <- 0 until 12) {
+    test("CountOneOnEach" + bitCount) {
+      SimConfig.noOptimisation
+        .compile(new Component {
+          val input = in Bits (bitCount bits)
+          val output = out Vec (CountOneOnEach(input))
+        })
+        .doSim(seed = 42) { dut =>
+          for (_ <- 0 until 100 + (1 << bitCount) * 4) {
+            dut.input.randomize()
+            sleep(1)
+            val input = dut.input.toBigInt
+            for (i <- 0 until bitCount; mask = ((1 << i + 1) - 1))
+              assert(dut.output(i).toInt === (input & mask).bitCount)
+          }
+        }
+    }
   }
 }
